% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_relative_abundance_assay.R
\name{add_relative_abundance_assay}
\alias{add_relative_abundance_assay}
\title{Add Relative Abundance Assay for a Single Taxonomic Level}
\usage{
add_relative_abundance_assay(qf, assay_name = "phylum")
}
\arguments{
\item{qf}{A QFeatures object.}

\item{assay_name}{Name of the taxonomic assay to convert. One of: "domain",
"kingdom", "phylum", "class", "order", "family", "genus", "species".}
}
\value{
A QFeatures object with an additional \code{{assay_name}_rel_abundance} assay.
}
\description{
Adds a new assay containing relative abundance values (as percentages) for
one taxonomic level in a QFeatures object.
}
