% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_relative_abundance_assay.R
\name{add_relative_abundance_assays}
\alias{add_relative_abundance_assays}
\title{Add Relative Abundance Assays for Multiple Taxonomic Levels}
\usage{
add_relative_abundance_assays(
  qf,
  assay_names = c("domain", "kingdom", "phylum", "class", "order", "family", "genus",
    "species")
)
}
\arguments{
\item{qf}{A QFeatures object.}

\item{assay_names}{Character vector of assay names to convert. Defaults to all
standard taxonomic levels that are present in \code{qf}.}
}
\value{
A QFeatures object with additional \code{{assay_name}_rel_abundance} assays.
}
\description{
Wrapper around \code{\link{add_relative_abundance_assay}} that converts all
standard taxonomic-level assays present in a QFeatures object to relative
abundance (percentage) assays.
}
\examples{
# qf_with_rel_abundance <- add_relative_abundance_assays(qfeatures_obj)
# qf_with_rel_abundance <- add_relative_abundance_assays(qfeatures_obj,
#   assay_names = c("genus", "species"))
}
