% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_percent_proteins_detected.R
\name{calc_percent_proteins_detected}
\alias{calc_percent_proteins_detected}
\title{calc_percent_proteins_detected}
\usage{
calc_percent_proteins_detected(
  conduit_obj,
  type = "multiple_proteins_in_group"
)
}
\arguments{
\item{conduit_obj}{tibble with columns named...}

\item{type}{multiple_proteins_in_group", "one_protein_in_group}
}
\description{
calculate the percent of proteins detected.
}
