% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_service.R
\name{check_api_service}
\alias{check_api_service}
\title{Title}
\usage{
check_api_service(api_url = "https://rest.uniprot.org/proteomes/UP000005640")
}
\arguments{
\item{api_url}{}
}
\description{
Title
}
