% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_proteins_per_taxon.R
\name{count_proteins_per_taxon}
\alias{count_proteins_per_taxon}
\title{count_proteins_per_taxon}
\usage{
count_proteins_per_taxon(tibble)
}
\arguments{
\item{tibble}{a tibble with columns organism_type, domain, kingdom,
phylum, class, order, family, genus, and species}
}
\description{
count the number of proteins per taxon
}
