% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hierarchy_taxa_count.R
\name{create_hierarchy_taxa_count}
\alias{create_hierarchy_taxa_count}
\title{create_hierachy_taxa_count}
\usage{
create_hierarchy_taxa_count(
  data,
  levels = c("domain", "kingdom", "phylum", "class", "order", "family", "genus",
    "species")
)
}
\arguments{
\item{levels}{}
}
\description{
create_hierachy_taxa_count
}
