% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_probiomeconduit_obj.R
\name{create_probiomeconduit_obj}
\alias{create_probiomeconduit_obj}
\title{create_probiomeconduit_obj
create a pbiomeconduit obj}
\usage{
create_probiomeconduit_obj(
  qfeature.rds,
  database_taxonomy.tsv,
  database_metrics.tsv,
  detected_protein_taxonomy.tsv,
  detected_protein_metrics.tsv
)
}
\arguments{
\item{detected_protein_metrics.tsv}{}
}
\description{
create_probiomeconduit_obj
create a pbiomeconduit obj
}
