% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_to_qfeatures.R
\name{diann_to_qfeatures}
\alias{diann_to_qfeatures}
\title{Convert DIANN Parquet file into a QFeatures object containing precursors,
peptides, and proteins, all with assay links.}
\usage{
diann_to_qfeatures(diann_parquet_fp)
}
\arguments{
\item{diann_parquet_fp}{}
}
\value{
a QFeatures object
}
\description{
Convert DIANN Parquet file into a QFeatures object containing precursors,
peptides, and proteins, all with assay links.
}
