% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cazyme_info.R
\name{get_cazyme_info}
\alias{get_cazyme_info}
\title{Getting Cazyme infromation for a vector of cazyme ids}
\usage{
get_cazyme_info(cazyme_ids)
}
\arguments{
\item{uniprot_annotated_data}{}
}
\description{
Getting Cazyme infromation for a vector of cazyme ids
}
