% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connected_features.R
\name{get_connected_features}
\alias{get_connected_features}
\title{get_connected_features}
\usage{
get_connected_features(qf, i = "protein_groups", am_name = "adjacencyMatrix")
}
\arguments{
\item{qf}{QFeatures object}

\item{i}{SE Name}

\item{id}{id that you would like to get connected features for}
}
\description{
get_connected_features
}
