% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxaids_from_species_names.R
\name{get_taxaids_from_species_names}
\alias{get_taxaids_from_species_names}
\title{Get NCBI taxonomy ids from species names}
\usage{
get_taxaids_from_species_names(species_names)
}
\arguments{
\item{species_names}{a charachter vector of NCBI taxonomy ids.}
}
\value{
a tibble of the species name and ncbi taxonomy id
}
\description{
Get the NCBI taxonomy ids corresponding to a vector of species names in a
convient tibble.
}
\examples{
get_taxaids_from_species_names(c("Homo sapiens", "Clostridium symbiosum"))
}
