% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxaid_from_species_name.R
\name{get_taxid_from_species_name}
\alias{get_taxid_from_species_name}
\title{Get NCBI taxa id corresponding to species name}
\usage{
get_taxid_from_species_name(organism)
}
\arguments{
\item{organism}{species name (string)}
}
\value{
NCBI taxonomy id if found, or NA and informative message.
}
\description{
This function will search a string of species names and return the NCBI
taxonomy id for that string. If a match isn't found, an informative message
and NA is returned. If more than one match is found, an informative message
and the first match is returned.
}
\examples{
get_taxaid_from_species_name("homo sapiens")
}
