% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_assay_to_species.R
\name{normalize_assay_to_species}
\alias{normalize_assay_to_species}
\title{Normalize assay abundance to species abundance}
\usage{
normalize_assay_to_species(
  qf,
  protein_assay = "protein_groups_log2",
  species_assay = "species_log2",
  species_col = "species",
  out_assay = "protein_groups_log2_species_normalized"
)
}
\arguments{
\item{qf}{QFeatures}

\item{protein_assay}{assay holding log2 protein counts}

\item{species_assay}{assay holding log2 species count}

\item{species_col}{column in rowData of protein_assay indicating species
the feature was derived from.}

\item{out_assay}{name of assay normalized to species abundance.}
}
\value{
The input \code{QFeatures} object with an additional assay (name given by
\code{out_assay}) containing protein abundances normalized to species abundance.
}
\description{
It may be adventageous to normalize protein intensities to total organism
abundance, paticularly when assessing a question of the type
. See https://peerj.com/preprints/2846/#.
}
\examples{
\dontrun{
qf <- normalize_assay_to_species(qf,
  protein_assay = "protein_groups_log2",
  species_assay = "species_log2",
  out_assay = "protein_groups_log2_species_normalized"
)
}
}
