% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_limma_analysis.R
\name{perform_limma_analysis}
\alias{perform_limma_analysis}
\title{perform_limma_analysis}
\usage{
perform_limma_analysis(qf, assay_name, formula, contrast)
}
\arguments{
\item{assay_name}{Name of the assay to analyze}

\item{formula}{A formula used to build the design matrix from colData}

\item{contrast}{A character string specifying the contrast (e.g., "groupB - groupA")}

\item{qf_object}{A QFeatures object}
}
\value{
A list containing:
\item{top_table}{The result of limma::topTable()}
\item{design}{The design matrix}
\item{coefficients}{Estimated model coefficients}
\item{model_terms}{Names of the model terms (design matrix columns)}
}
\description{
Performs limma differential expression analysis on a specified assay within a QFeatures object.
}
