% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colData_distribution.R
\name{plot_colData_distribution}
\alias{plot_colData_distribution}
\title{plot_colData_distribution'}
\usage{
plot_colData_distribution(conduit_obj, x_axis)
}
\arguments{
\item{x_axis}{}
}
\description{
plot the distribution of a variable present in the colData.
}
