% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conduit_metric_diann.R
\name{plot_conduit_metric_diann}
\alias{plot_conduit_metric_diann}
\title{Plot Diann Sample Level Metrics for Conduit Object}
\usage{
plot_conduit_metric_diann(conduit_obj, column_name = "MS1.Signal")
}
\arguments{
\item{conduit_obj}{conduit object}

\item{column_name}{column name to plot}
}
\description{
Plot Diann Sample Level Metrics for Conduit Object
}
