% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confusion_matrix.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{plot a confusion matrix}
\usage{
plot_confusion_matrix(predict_classification_list)
}
\arguments{
\item{predict_classification_list}{}
}
\description{
plot a confusion matrix
}
