% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_importance.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{plot feature importance}
\usage{
plot_feature_importance(predict_classification_list, start = 1, end = 10)
}
\arguments{
\item{end}{}
}
\description{
plot feature importance
}
