% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features_per_sample.R
\name{plot_features_per_sample}
\alias{plot_features_per_sample}
\title{Title}
\usage{
plot_features_per_sample(qf, assay = "protein_groups")
}
\arguments{
\item{assay}{}
}
\description{
Title
}
