% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  qf,
  assay_name,
  col_color_variables = NULL,
  row_color_variables = NULL,
  scale = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Easy heatmap plots.
}
