% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_precision_recall.R
\name{plot_precision_recall}
\alias{plot_precision_recall}
\title{plot precision recall curves from predicted classifications}
\usage{
plot_precision_recall(predict_classification_list, data_set = "test")
}
\arguments{
\item{data_set}{}
}
\description{
plot precision recall curves from predicted classifications
}
