% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_abundance.R
\name{plot_relative_abundance}
\alias{plot_relative_abundance}
\title{plot_relative_abundance
Plot the relative abundance for all samples.}
\usage{
plot_relative_abundance(qf, assay_name, facet_formula)
}
\arguments{
\item{qf}{q features object}

\item{assay_name}{assay to get relative abundance for}

\item{facet_formula}{formula to specify how the plot is faceted}
}
\description{
plot_relative_abundance
Plot the relative abundance for all samples.
}
