% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_selected_features.R
\name{plot_selected_features}
\alias{plot_selected_features}
\title{plot_selected_features}
\usage{
plot_selected_features(
  qf,
  assay_name,
  features,
  x_axis,
  facet_formula = NULL,
  color_by = NULL,
  shape = NULL
)
}
\arguments{
\item{shape}{}
}
\description{
plot the selected features
}
