% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano()}
\usage{
plot_volcano(
  limma_stats,
  facet_formula = NULL,
  color_by = NULL,
  pval_threshold = 0.05
)
}
\arguments{
\item{pval_threshold}{}
}
\description{
plot_volcano()
}
