% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduit.R
\name{show,conduit-method}
\alias{show,conduit-method}
\title{Show Method for Conduit Object}
\usage{
\S4method{show}{conduit}(object)
}
\arguments{
\item{object}{A conduit object to summarize}
}
\description{
Displays a summary of the conduit object's contents, including the number of
assays and availability of various data components.
}
\examples{
# Display summary of a conduit object:
# show(conduit_obj)
}
