% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_conduit.R
\name{tidy_conduit}
\alias{tidy_conduit}
\title{tidy_conduit}
\usage{
tidy_conduit(qf, assay_name)
}
\arguments{
\item{assay_name}{}
}
\description{
tidy data from a QFeatures object.This function combines,
assay data, colData, and rowData into a tbl_df in long format.
}
