% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_uniprot_accession_ids.R
\name{validate_uniprot_accession_ids}
\alias{validate_uniprot_accession_ids}
\title{validate_uniprot_accession_ids}
\usage{
validate_uniprot_accession_ids(uniprot_ids)
}
\arguments{
\item{uniprot_ids}{}
}
\description{
validates uniprot accession ids using regular expression. Filters out ones
that don't match.
}
