% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_log_imputed_norm_assays.R
\name{add_log_imputed_norm_assays}
\alias{add_log_imputed_norm_assays}
\title{add_log_imputed_norm_assays}
\usage{
add_log_imputed_norm_assays(
  qf,
  base = 2,
  impute_method = "min",
  norm_method = "none"
)
}
\arguments{
\item{norm_method}{}
}
\description{
add_log_imputed_norm_assays
}
