% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_relative_abundance_assays.R
\name{add_relative_abundance_assays}
\alias{add_relative_abundance_assays}
\title{add_relative_abundance_assays
add an assay containing relative abundance (as a proportion) to a qfeatures
object.}
\usage{
add_relative_abundance_assays(
  qf,
  assay_names = c("domain", "kingdom", "phylum", "class", "order", "family", "genus",
    "species")
)
}
\arguments{
\item{assay_name}{}
}
\description{
add_relative_abundance_assays
add an assay containing relative abundance (as a proportion) to a qfeatures
object.
}
