% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rowdata_to_limma_results.R
\name{add_rowdata_to_limma_results}
\alias{add_rowdata_to_limma_results}
\title{add_rowdata_to_limma_results}
\usage{
add_rowdata_to_limma_results(limma_stats, qf, assay_name)
}
\arguments{
\item{assay_name}{}
}
\description{
Add rowdata to the statistics of limma data.
}
