% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_uniprot_ids.R
\name{annotate_uniprot_ids}
\alias{annotate_uniprot_ids}
\title{annotate_uniprot_ids}
\usage{
annotate_uniprot_ids(
  uniprot_ids,
  columns = NULL,
  batch_size = 150,
  parallel = TRUE
)
}
\arguments{
\item{parallel}{}
}
\description{
annotate_uniprot_ids
}
