% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_relative_abundance.R
\name{calc_relative_abundance}
\alias{calc_relative_abundance}
\title{calc_relative_abundance}
\usage{
calc_relative_abundance(qf, assay_name)
}
\arguments{
\item{assay_name}{}
}
\description{
calculate the relative abundance of an QFeatures assay.
}
