% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_fasta_files.R
\name{concatenate_fasta_files}
\alias{concatenate_fasta_files}
\title{concatenate_fasta_files
concatenate fasta files.}
\usage{
concatenate_fasta_files(fasta_dir, destination_fp)
}
\arguments{
\item{output_file}{}
}
\description{
concatenate_fasta_files
concatenate fasta files.
}
