% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduit.R
\docType{class}
\name{conduit-class}
\alias{conduit-class}
\title{conduit Class}
\description{
A class to store QFeatures and additional data about the database used for
detection, as well as the detected proteins.
}
\section{Slots}{

\describe{
\item{\code{qfeatures}}{A QFeatures object containing proteomics data.}

\item{\code{experiment_metadata}}{A tibble containing the experimental metadata.}

\item{\code{database_taxonomy}}{A tibble containing taxonomy information for the database.}

\item{\code{database_metrics}}{A tibble containing metrics related to the database.}

\item{\code{detected_protein_taxonomy}}{A tibble containing taxonomy information for detected proteins.}

\item{\code{detected_protein_metrics}}{A tibble containing metrics for detected proteins.}
}}

