% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_conduit_obj.R
\name{create_conduit_obj}
\alias{create_conduit_obj}
\title{create_conduit_obj
create a conduit object.}
\usage{
create_conduit_obj(
  qfeature.rds,
  database_taxonomy.tsv,
  database_metrics.tsv,
  detected_protein_taxonomy.tsv,
  detected_protein_metrics.tsv
)
}
\arguments{
\item{detected_protein_metrics.tsv}{}
}
\description{
create_conduit_obj
create a conduit object.
}
