% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fasta_from_proteome_ids.R
\name{download_fasta_from_proteome_ids}
\alias{download_fasta_from_proteome_ids}
\title{Title}
\usage{
download_fasta_from_proteome_ids(
  proteome_ids,
  parallel = FALSE,
  proteome_id_destination_fp = paste0(getwd(), "/", Sys.Date(), ".txt"),
  fasta_destination_fp = paste0(getwd(), "/", Sys.Date(), ".fasta")
)
}
\arguments{
\item{fasta_destination_fp}{}
}
\description{
Title
}
