% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fasta_info.R
\name{extract_fasta_info}
\alias{extract_fasta_info}
\title{Extract Information from FASTA Files}
\usage{
extract_fasta_info(fasta_file)
}
\arguments{
\item{fasta_file}{Character string specifying the path to a FASTA file.
The file should contain protein sequences with headers in UniProt format,
which typically include:
\itemize{
\item Protein ID (e.g., "sp|P12345|PROT_NAME")
\item Organism name (OS=...)
\item Organism ID (OX=...)
}}
}
\value{
A data frame containing:
\itemize{
\item protein_id: UniProt protein accession ID
\item organism_name: Scientific name of the organism
\item organism_id: NCBI taxonomy ID of the organism
\item sequence: Amino acid sequence of the protein
}
Missing values (NA) are used for any information that cannot be extracted
from the headers.
}
\description{
Parses a FASTA file and extracts key information including protein IDs,
organism names, organism IDs, and protein sequences. This function is
particularly useful for processing UniProt FASTA files and extracting
metadata from the headers.
}
\examples{
\dontrun{
protein_info <- extract_fasta_info("uniprot_proteins.fasta")
head(protein_info[, c("protein_id", "organism_name", "organism_id")])
# Use for taxonomy matrices or mapping proteins to organisms
}
# Expects UniProt-style headers (OS=, OX=); other formats may not parse.

}
