% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_possible_contrast_terms.R
\name{find_possible_contrast_terms}
\alias{find_possible_contrast_terms}
\title{find_possible_contrast_terms
Helper function to assist the user in finding contrast terms that are
possible given the specified formula.}
\usage{
find_possible_contrast_terms(qf, assay_name, formula)
}
\arguments{
\item{formula}{}
}
\description{
find_possible_contrast_terms
Helper function to assist the user in finding contrast terms that are
possible given the specified formula.
}
