% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations_from_uniprot.R
\name{get_annotations_from_uniprot}
\alias{get_annotations_from_uniprot}
\title{get_annotations_from_uniprot}
\usage{
get_annotations_from_uniprot(
  uniprot_ids,
  columns =
    "accession,go,xref_kegg,xref_eggnog,cc_subcellular_location,cc_tissue_specificity,xref_cazy,xref_pfam,xref_interpro"
)
}
\arguments{
\item{pg_matrix_fp}{pg_matrix output of diann}

\item{outdir}{out directory}
}
\value{
tsv at output directory location
}
\description{
given the pg_matrix output of diann, assign go terms to each protein group
and summarise.
}
\examples{
insulin <- get_annotations_from_uniprot("P01308")
t <- get_annotations_from_uniprot("P47340")
}
