% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_better_proteome_ids.R
\name{get_better_proteome_ids}
\alias{get_better_proteome_ids}
\title{Select the best UniProt proteome for a given taxonomic identifier.}
\usage{
get_better_proteome_ids(proteome_id_df, parallel = FALSE)
}
\arguments{
\item{proteome_id_df}{Data frame containing annotations produced by the
\code{get_proteome_ids()} function.}

\item{parallel}{Logical; whether to perform selection in parallel (default: FALSE).}
}
\value{
A data frame with an additional column \code{selected_proteome_id} indicating
the chosen proteome for each taxonomic identifier.
}
\description{
Sometimes, the proteome IDs returned for a specific NCBI taxonomic identifier
are not available in UniProtKB. To maximize the information obtained from
these identifiers, this function applies the following selection logic:
}
\details{
\enumerate{
\item \strong{Strain-level taxonomic IDs:}
\itemize{
\item If a proteome exists that is classified as "Representative proteome",
"Reference and representative proteome", or "Other proteome", it will be
used.
\item If no such proteome exists, the function moves up to the species level
and selects proteomes in the following order of preference:
"Representative proteome", "Reference and representative proteome",
"Other proteome", "Redundant proteome", and "Excluded proteome".
}
\item \strong{Species-level taxonomic IDs:}
\itemize{
\item If a redundant proteome exists, it will be used as the selected
proteome.
}
}

The first proteome meeting these criteria will be returned as \code{selected_proteome_id}.
}
\examples{
# Data format compatible with get_better_proteome_ids()
proteome_id_df <- tibble::tibble(
  proteome_id = "UP000027861",
  organism_id = 1339341,
  organism = "Parabacteroides distasonis str. 3776 Po2 i",
  protein_count = 4580,
  proteome_type = "Redundant proteome",
  redundant_to = "UP000027850",
  genome_assembly_id = "GCA_000699745.1",
  genome_assembly_level = "full",
  annotation_score = NA)

# Example function call
better_proteome_ids <- get_better_proteome_ids(proteome_id_df)
}
