% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_file.R
\name{get_fasta_file}
\alias{get_fasta_file}
\title{get_fasta_file
download proteome file from uniprot.}
\usage{
get_fasta_file(proteome_id, fasta_dir = getwd())
}
\arguments{
\item{proteome_id}{proteome id}

\item{fasta_fp}{file path to write to}
}
\description{
get_fasta_file
download proteome file from uniprot.
}
