% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteome_id_from_organism_id.R
\name{get_proteome_id_from_organism_id}
\alias{get_proteome_id_from_organism_id}
\title{get_batch_of_proteome_ids}
\usage{
get_proteome_id_from_organism_id(id)
}
\arguments{
\item{base_url}{}
}
\description{
get a batch of proteome ids from organism ids
}
\details{
Note this can only handle
}
