% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteome_ids_from_organism_ids.R
\name{get_proteome_ids_from_organism_ids}
\alias{get_proteome_ids_from_organism_ids}
\title{get_proteome_ids_from_organism_ids}
\usage{
get_proteome_ids_from_organism_ids(organism_ids, parallel = TRUE)
}
\arguments{
\item{parallel}{}
}
\description{
get_proteome_ids_from_organism_ids
}
