% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteome_taxids_and_types.R
\name{get_proteome_taxids_and_types}
\alias{get_proteome_taxids_and_types}
\title{Title}
\usage{
get_proteome_taxids_and_types(proteome_ids, parallel = FALSE)
}
\arguments{
\item{proteome_ids}{}
}
\description{
Title
}
