% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduit.R
\name{initialize,conduit-method}
\alias{initialize,conduit-method}
\title{Initialize a Conduit object}
\usage{
\S4method{initialize}{conduit}(
  .Object,
  QFeatures,
  combined_metrics = NULL,
  database_protein_taxonomy = NULL,
  detected_protein_taxonomy = NULL
)
}
\arguments{
\item{QFeatures}{A QFeatures object.}

\item{combined_metrics}{A tibble containing proteins and their assigned taxonomy, as used in database}

\item{database_protein_taxonomy}{A tibble containing the database metrics.}

\item{detected_protein_taxonomy}{A tibble containing the detected protein taxonomy.}
}
\value{
A Conduit object.
}
\description{
Initialize a Conduit object
}
