% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduit.R
\name{initialize,conduit-method}
\alias{initialize,conduit-method}
\title{Initialize a Conduit object}
\usage{
\S4method{initialize}{conduit}(
  .Object,
  QFeatures,
  database_taxonomy = NULL,
  database_metrics = NULL,
  detected_protein_taxonomy = NULL,
  detected_protein_metrics = NULL
)
}
\arguments{
\item{database_taxonomy}{A tibble containing the database taxonomy.}

\item{database_metrics}{A tibble containing the database metrics.}

\item{detected_protein_taxonomy}{A tibble containing the detected protein taxonomy.}

\item{detected_protein_metrics}{A tibble containing the detected protein metrics.}

\item{qf}{A QFeatures object.}

\item{experiment_metadata}{A tibble containing the experiment metadata.}
}
\value{
A Conduit object.
}
\description{
Initialize a Conduit object
}
