% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_gsea.R
\name{perform_gsea}
\alias{perform_gsea}
\title{Title}
\usage{
perform_gsea(
  limma_stats,
  conduit,
  annotation_type = "go",
  ranking_column = "logFC"
)
}
\arguments{
\item{ranking_column}{}
}
\description{
Title
}
