% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_ora.R
\name{perform_ora}
\alias{perform_ora}
\title{perform_ora}
\usage{
perform_ora(
  limma_stats,
  direction = "up",
  conduit,
  annotation_type = "species",
  adj_pval_threshold = 0.05,
  logFC_threshold = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
perform_ora
}
