% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biplot.R
\name{plot_biplot}
\alias{plot_biplot}
\title{plot_biplot}
\usage{
plot_biplot(
  qf,
  assay_name,
  color = NULL,
  shape = NULL,
  removeVar = 0.1,
  legendPosition = "right",
  facet_formula = NULL
)
}
\arguments{
\item{legendPosition}{}
}
\description{
Plot a PCA biplot of an assay in QFeatures object.
}
