% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colData_distribution.R
\name{plot_colData_distribution}
\alias{plot_colData_distribution}
\title{Plot Distribution of Column Metadata}
\usage{
plot_colData_distribution(conduit_obj, x_axis)
}
\arguments{
\item{conduit_obj}{A conduit object containing the QFeatures data}

\item{x_axis}{Character string specifying which column from colData to plot}
}
\value{
A ggplot object containing:
\itemize{
\item Histogram showing the distribution of the specified variable
\item X-axis showing the variable values
\item Y-axis showing the count/frequency
}
}
\description{
Creates a histogram showing the distribution of a variable from the column metadata
(colData) of a conduit object. This is useful for exploring the distribution of
sample characteristics and experimental variables.
}
\examples{
# Plot distribution of a categorical variable:
# plot_colData_distribution(conduit_obj, "group")

# Plot distribution of a numeric variable:
# plot_colData_distribution(conduit_obj, "concentration")

# Note: For numeric variables, you might want to use geom_density() instead
# of the default geom_histogram() by modifying the returned plot
}
