% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaply.R
\name{plot_heatmaply}
\alias{plot_heatmaply}
\title{plot_heatmaply}
\usage{
plot_heatmaply(
  qf,
  assay_name,
  col_color_variables = NULL,
  row_color_variables = NULL,
  scale = "column",
  hover_col_variables = NULL,
  hover_row_variables = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
plot_heatmaply
}
