% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kegg_pathway.R
\name{plot_kegg_pathway}
\alias{plot_kegg_pathway}
\title{Plot Data on KEGG Pathway Map}
\usage{
plot_kegg_pathway(stats_results, kegg_pathway_id, fill_by = logFC)
}
\arguments{
\item{stats_results}{some sort of statistics results (intended to be the
toptable produced by the perfom_limma_stats function). Must have a column
named xref_kegg containing the kegg id corresponding to each protein.}

\item{kegg_pathway_id}{Character string specifying the KEGG pathway ID to
plot}

\item{fill_by}{the metric to fill the color of the boxes by.}
}
\description{
Visualizes protein or gene expression data on a KEGG pathway map using the
ggkegg package.
}
\details{
This function creates a pathway diagram where nodes (proteins/genes) are colored according
to their expression values or other quantitative data.
}
\examples{
# Note: KEGG pathway IDs can be found at https://www.genome.jp/kegg/pathway.html
}
