% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_percent_detected_taxa_tree.R
\name{plot_percent_detected_taxa_tree}
\alias{plot_percent_detected_taxa_tree}
\title{plot_percent_detected_taxa_tree}
\usage{
plot_percent_detected_taxa_tree(
  conduit_obj,
  type = "multiple_proteins_in_group",
  ...
)
}
\arguments{
\item{conduit_obj}{a conduit object}

\item{...}{any argument taken by metacoder::heat_tree}
}
\description{
plot_percent_detected_taxa_tree
}
