% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_cor_heatmap.R
\name{plot_sample_cor_heatmap}
\alias{plot_sample_cor_heatmap}
\title{plot_sample_cor_heatmap}
\usage{
plot_sample_cor_heatmap(qf, assay_name, sample_annotation_variables)
}
\arguments{
\item{sample_annotation_variables}{}
}
\description{
plot the pearson correlation.
}
