% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sunburst.R
\name{plot_sunburst}
\alias{plot_sunburst}
\title{plot_sunburst}
\usage{
plot_sunburst(taxonomy)
}
\arguments{
\item{taxonomy}{tibble with taxonomy information. Must have columns
superkingdom,kindom,phylum,class,order,family,genus,species.}
}
\description{
Create a sunburst plot.
}
