% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa_tree.R
\name{plot_taxa_tree}
\alias{plot_taxa_tree}
\title{plot_taxa_tree}
\usage{
plot_taxa_tree(
  taxonomy_data,
  filter_taxa_rank = "species",
  node_size = "n_obs",
  ...
)
}
\arguments{
\item{taxonomy_data}{taxonomy data as a tibble. Each observation is a row,
columns are.}

\item{filter_taxa_rank}{what rank to filter taxa by.}

\item{node_size}{what variable to color node size by.}
}
\description{
plot a taxa tree using taxonomy data.
}
