% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_regression.R
\name{predict_regression}
\alias{predict_regression}
\title{predict regession}
\usage{
predict_regression(
  qf,
  assay_name,
  outcome,
  train_percent = 70,
  model_type = "lasso_regression",
  v = 5,
  grid_size = 20
)
}
\arguments{
\item{grid_size}{}
}
\description{
Predict numeric outcomes
}
