% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_qfeature.R
\name{prepare_qfeature}
\alias{prepare_qfeature}
\title{prepare_qfeature}
\usage{
prepare_qfeature(
  sample_annotation_fp,
  vector_of_matrix_fps,
  vector_of_matrices_names = NULL
)
}
\arguments{
\item{sample_annotation_fp}{filepath of sample annotation file path}

\item{vector_of_matrix_fps}{vector of filepath}

\item{vector_of_matrices_names}{vector of names to call each matrices.
Optional, by default the file name - the extension will be used.}
}
\description{
assemble qfeature object from a list of any number of matrices corresponding
to the outputs of the process.
}
