% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduit.R
\name{show,conduit-method}
\alias{show,conduit-method}
\title{Show method for Conduit object}
\usage{
\S4method{show}{conduit}(object)
}
\arguments{
\item{object}{A Conduit object.}
}
\description{
Show method for Conduit object
}
